plan out all the things to test
