+++ packages/ws-litefs/src/__tests__/LiteFSDBFactory.test.ts
@@ -1,0 +1,7 @@
+import { test, expect } from "vitest";
+
+test("forwards db create event to primary when follower", () => {});
+
+test("awaits txid to catch up, after db creation, when follower", () => {});
+
+test("No waiting for tx when primary", () => {});
