+++ packages/ws-litefs/src/__tests__/util.test.ts
@@ -1,0 +1,11 @@
+import { test, expect } from "vitest";
+
+test("reading txid", () => {});
+
+test("wait until a given txid", () => {});
+
+test("throwing inside waitUntil correctly cleans up listeners", () => {});
+
+test("read the primary file when it does not exist", () => {});
+
+test("read the primary file when it does exist", () => {});
