+++ packages/ws-litefs/src/config.ts
@@ -1,0 +1,2 @@
+export let port = 9000;
+export let primaryFilePath = "/var/lib/litefs/.primary";
