+++ packages/ws-litefs/src/internal/__tests__/PrimaryConnection.test.ts
@@ -1,0 +1,11 @@
+import { test, expect } from "vitest";
+
+test("Returns that it is the primary if the primary file is missing", () => {});
+
+test("Upgrades self to primary if the primary file exists at construction", () => {});
+
+test("Upgrades self to primary if the primary file is created post-construction", () => {});
+
+test("Downgrades self to follower if the primary file is removed post-construction", () => {});
+
+test("Swaps connection when primary changes", () => {});
