+++ packages/ws-litefs/src/internal/__tests__/PrimarySocket.test.ts
@@ -1,0 +1,15 @@
+import { test, expect } from "vitest";
+
+test("send create db does not resolve until receiving a response", () => {});
+
+test("we throw when receiving an unknown reqid", () => {});
+
+test("repsonses are correctly paired with requests", () => {});
+
+test("pending requests are removed when receiving a response", () => {});
+
+test("pending requests are removed on error", () => {});
+
+test("pending requests are removed on close", () => {});
+
+test("socket closed after 5-6 seconds of missed heartbeats", () => {});
